# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "micro de servicios.",
        "title": "medfri-service",
        "contact": {
            "name": "Soporte de API",
            "url": "http://www.soporte-api.com",
            "email": "soporte@api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9020",
    "basePath": "/medfri-service",
    "paths": {
        "/servicio/{prefijo}": {
            "get": {
                "description": "obtener servicio.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "servicio"
                ],
                "summary": "obtener prefijo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "prefijo o abreviatura",
                        "name": "prefijo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Error en el cuerpo de la solicitud",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ingresa \"Bearer {token}\" para autenticar.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}